Per-request headers, cache keys from an expression, GraphQL answers shaped by the declared type, and the editor running validate's checks

+++ internal/flow/cache_codec_test.go
@@ -149,8 +149,10 @@ func TestDecodeCacheValue_TolerantBase64(t *testing.T) {
 			if err != nil {
 				t.Fatalf("decode: %v", err)
 			}
+			// An integral number reads back as an int64 since the decode
+			// stopped flattening every number to float64.
 			m, ok := got.(map[string]interface{})
-			if !ok || m["a"] != float64(1) {
+			if !ok || m["a"] != int64(1) {
 				t.Errorf("got %#v", got)
 			}
 		})
